refactor: split refresh into its own logic path
