+++ src/handlers/logic/auth.rs
@@ -1,3 +1,5 @@
 mod login;
+mod refresh;
 
 pub use login::*;
+pub use refresh::*;
